client: refactor(shared/ui/footer): reorganize footer component according to FSD
