--- client/src/app/shared/ui/footer/footer.component.html
@@ -1,3 +1,0 @@
-<div class="footer">
-  <p>EnglishAiEasy LLC &copy; 2023-{{ dateNow() }}</p>
-</div>
