+++ client/src/app/shared/ui/footer/index.ts
@@ -1,0 +1,1 @@
+export { FooterComponent } from './ui/footer.component';
