+++ client/src/app/shared/ui/footer/ui/footer.component.html
@@ -1,0 +1,3 @@
+<div class="footer">
+  <p>EnglishAiEasy LLC &copy; 2023-{{ dateNow() }} - Online: {{ network.online$ | async }}</p>
+</div>
