+++ client/src/app/shared/ui/footer/ui/footer.component.spec.ts
