+++ client/src/app/shared/ui/footer/ui/footer.component.ts
@@ -1,14 +1,18 @@
-import { Component, OnInit, signal } from '@angular/core';
+import { NetworkService } from '@shared/infrastructure/network.service';
+
+import { AsyncPipe } from '@angular/common';
+import { Component, inject, OnInit, signal } from '@angular/core';
 
 @Component({
   selector: 'app-footer',
   standalone: true,
-  imports: [],
+  imports: [AsyncPipe],
   templateUrl: './footer.component.html',
   styleUrl: './footer.component.scss',
 })
 export class FooterComponent implements OnInit {
   dateNow = signal<string>('');
+  network = inject(NetworkService);
 
   ngOnInit(): void {
     this.dateNow.set(new Date().getFullYear().toString());
