+++ client/src/app/shared/ui/index.ts
@@ -1,4 +1,4 @@
 export { DataTableComponent } from './data-table/data-table.component';
 export { ErrorMessageComponent } from './error-message/error-message.component';
-export { FooterComponent } from './footer/footer.component';
+export { FooterComponent } from './footer';
 export { SnackBarComponent } from './snack-bar/snack-bar.component';
